requirements.txt
requests>=2.31.0
PyJWT>=2.8.0
cryptography>=41.0.0

